Fixing decimal support on NEO NEP5 and Assets Transactions transactions.
Fixing decimal support on get balance for NEO NEP5 Contracts.

+++ NeoFlux/Controllers/TransactionController.cs
@@ -21,7 +21,7 @@ namespace NeoFlux.Controllers
             var toAddress = jsonData.GetValue("recipientAddress").Value<string>();
             
             /** Get the type of currency */
-            var amount = jsonData.GetValue("amount").Value<int>();
+            var amount = jsonData.GetValue("amount").Value<decimal>();
             var asset = jsonData.GetValue("asset").Value<string>();
             
             /* Get owner data from private key */
